requirements.txt
Django==4.1.1
Pillow==9.2.0
psycopg2-binary==2.8.6
python-dotenv==0.21.0
sqlparse==0.4.2
djangorestframework==3.12.4